third_party/blink/web_tests/external/wpt/domxpath/resolver-callback-interface-cross-realm.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] XPathNSResolver is cross-realm plain object without 'lookupNamespaceURI' property
  assert_equals: expected "object" but got "undefined"
[FAIL] XPathNSResolver is cross-realm plain object with non-callable 'lookupNamespaceURI' property
  assert_equals: expected "object" but got "undefined"
[FAIL] XPathNSResolver is cross-realm non-callable revoked Proxy
  assert_equals: expected "object" but got "undefined"
[FAIL] XPathNSResolver is cross-realm callable revoked Proxy
  assert_equals: expected function "function TypeError() { [native code] }" but got function "function TypeError() { [native code] }"
[FAIL] XPathNSResolver is cross-realm plain object with revoked Proxy as 'lookupNamespaceURI' property
  assert_equals: expected function "function TypeError() { [native code] }" but got function "function TypeError() { [native code] }"
Harness: the test ran to completion.
